+++ shuup/admin/modules/products/validators.py
@@ -12,7 +12,8 @@ from shuup.core.models import Shop, ShopProduct, Supplier
 
 
 class AdminProductValidator:
-    """Base class for validating products."""
+    """ Base class for validating products. """
+    ordering = 0
 
     def get_validation_issues(
         shop_product: ShopProduct, shop: Shop, user, supplier: Supplier = None